+++ changelog.md
@@ -8,6 +8,7 @@
 * Remove redundant APPDATA_DIR loading in util (#665, #666)
 * Fix theme breakage caused by a promoted widget (#669, #670)
 * Fix random/nomads faction shown in replay info (#676)
+* Fix lobby name change not showing in Play (#690) 
 
 Contributors:
   - Wesmania